Split most of the implementation into a "core" library
This arranges to build the "dy-*.c" sources into a "libdinghycore" which
can be used as a basis for implementing other WPE-based launchers other
than "dinghy". This can be useful e.g. when one needs to customize
creation of the WebKitWebView to use a particular WPE backend and/or do
additional steps for WPE to work, without having the reimplement a
complete launcher.