Allow building with WebKitGTK+ instead of WPE
Both expose GLib-based APIs, which are so similar that they are almost
interchangeable. The main difference is that in WebKitGTK+ WebKitWebView
is a widget, and therefore ultimately it needs to be contained into some
kind of GtkWindow. For convenience, DyLauncher will inherit from
GtkApplication (instead of plain GApplication) when building against
WebKitGTK+.