DyLauncher: Add a "home-uri" property
This is the URI which gets loaded once the main instance is activated.
An application can set it e.g. from a GApplication::local-command-line
signal handler, or after instantiating DyLauncher. Once the web view is
created, the home URI will be loaded.

Setting the property afterwards, when the web view exists already, causes
the specified URI to be loaded into the web view. Remotely opening an URI
(over D-Bus) will also set the "home-uri" property and load it in the web
view.