Add a DRM/KMS mode monitor which implements DyModeMonitor
The new DyDrmModeMonitor watches udev events on a DRM/KMS device node
(typically /dev/dri/card0) and uses libdrm to query the status of the
CRTC/framebuffer associated with it on changes.

Currently, this mode monitor has a few shortcomings:

- Changes are only detected when the configuration of the DRM device
  pipeline (CRTCs, encoders, connectors) change: resolution changes
  started from userspace will go unadverted.

- When changes are detected, multiple uevents may be received. In order
  to throttle the amount of changes quickly detected, some hysteresis is
  used and DRM device status is read after 500ms without receiving new
  events.

- Monitoring multiple DRM devices is possible by creating multiple
  instances of the monitor, but untested. Each monitor will create a
  libudev monitor, which is unoptimal. In the future a singleton
  might be used, with the events being dispatched to the corresponding
  monitor object.

That being said, it works pretty well and may be useful in a number of
cases where the resolution chages happen externally in a way that the
DRM device notices them, e.g. an HDMI output which, when the output
resolution changes, reconfigures the CRTC and/or connector of the
corresponding DRM device.