Factor out URI command line guessing logic and use it in dinghyctl
This moves out the code which figures out whether a string is a local
file, and whether an URI needs a scheme prefix to its own function,
changes dinghy's main() to use this new function, and modifies dinghyctl
so the "open" subcommand uses it as well. This way both the programs use
the same logic for URI-like command line arguments.