dinghyctl: Fix segfault when using an unavailable command
The G_N_ELEMENTS() macro was returning one item too much due to having
an unneeded sentinel element in the array of command structures, which
would result in calling strcmp() with a NULL string when reaching the
sentinel. Given that all the lookups for command structures is done by
means of cmd_find_by_name() we can just remove the sentinel item from
the end of the array.