Fix handling of custom URI scheme handling in CLI arguments
For any URI that did not have either http:// or https:// or wasn't
a file the http:// prefix was being added inconditionally, which of
course breaks custom URI schemes. Instead of checking against a fixed
set of schemes, properly parse the URI and use the SoupURI APIs to
check and modify the URI fields, and re-construct the final URI which
gets passed to WebKit for loading.

Note that does not really add any dependency, as libsoup is needed
already by WebKitWPE (and WebKitGTK+).

Fixes #5