Implement restarting the WebProcess on failure
This brings in an utility function which will connect a signal handler
to WebKitWebView::web-process-crashed and restart the WebProcess up to
maximum number of tries over a window of time. The launcher is modified
to accept "--webprocess-failure=restart", in which case it will try at
most five times in a second. The number of retries and the time for now
are hardcoded.

Fixes #10