Support exporting remote control actions to system bus
This adds a built-time toggle DY_DBUS_SYSTEM_BUS, which will enable
exporting the org.gtk.Actions interface on the system bus. Note that
despite the name, this D-Bus interface does not require using GTK+
at all: the implementation which exports a GActionGroup over D-Bus
is readily available in GIO.

Note that the object path is always /com/igalia/Dinghy, regardless of
whether GTK+ support is built in or not, but the well-known D-Bus name
will be either the one manually configured at build time, or one of
com.igalia.Dinghy, or com.igalia.DinghyGtk depending on whether GTK+
support is enabled.

This is the first part of the work needed for completing #15