Do not hardcode D-Bus object path for the system bus
Factor out the “appid_to_object_path()” function into “dy-utils.[hc]”
and use it both in dinghyctl and DyLauncher to determine the default
object path to use on the system bus. The default behaviour renders the
same object path as used for the session bus, which is handy because it
makes “dinghyctl --system <command>” just work without needing to
manually specify the object path.