Make use of all cflags from pkg-config
CMake's FindPkgConfig is kinda weird. We need to ensure we pass on all
cflags, not just the includes. Otherwise we lose important stuff like
-DMESA_EGL_NO_X11_HEADERS from egl.pc, resulting in a broken
eglplatform.h on systems without libX11.

Note: we don't have to worry about this for linker flags, because we are
passing all LDFLAGS into target_link_libraries(). That's probably not
really the idiomatic CMake solution, but it seems like the best
approach. My attempt to improve on this ended in manual string
replacements, a bit too much work for what we know to be zero flags.