Un-hardcode usage of FDO the platform module
This adds a new --platform/-P command line flag which allows specifying
the name of the platform plug-in to be loaded. If none is specified,
a fall-back to the default WPE backend is used, which allows usage of
e.g. WPEBackend-rdk and others.

This also adds a few g_debug() calls sprinkled here and there. This makes
it easier to figure out which WPE backend is being used, and contributes
a bit to fix issue #1