CMake: Use add_library(foo MODULE ...) for plug-ins
This ensures that CMake passes any compiler and linker flags which might
be needed to ensure that a shared object can be loaded with dlopen().
Note that using MODULE results in shared objects which do not include a
version string in their name. This is fine, as in the future we may want
to store platform modules in a separate directort which does contain the
platform API version string (e.g. /usr/lib/cog-0.1/)