Skip error page when WebKit hands load to plug-in or media engine
This situation is easily triggered by loading an URL pointing directly
to a video, due to WebKit passing the responsibility of fetching data
to GStreamer. In that case, do not show an error page and return FALSE
in order to let WebKit continue with its fall-back handling, which will
let the plug-in or media engine do the loading.

Fixes #2