[PoC] Implement a "nil" platform, make platforms choose WPE backend
Also, change the FDO platform to set with an environment variable
that "libWPEBackend-fdo.so" is to be used as WPE backend library.

Note that this is a proof-of-concept and needs either a debug
build of libWPEBackend, or a patched version which enables it to
check the WPE_BACKEND_LIBRARY environment variable.

With this applied (and a suitable libWPEBackend) it is not needed
to have a libWPEBackend-default.so file, and the following command
lines will work fine provided that the FDO and RDK backends are
installed:

  % cog -P fdo $URI
  % cog -P nil:backend=libWPEBackend-rdk.so $URI

Note that when using the "nil" platform plug-in, if the "backend"
parameter is not specified, then "libWPEBackend-default.so" is still
used. On the other hand, the "fdo" platform plug-in forces usage of
"libWPEBackend-fdo-0.1.so" because it cannot work with anything else.

This is proof-of-concept implementation of #11