Provide command line options for WebKitSettings properties
This adds code to introspect properties of a given GObjectClass
and obtain a set of GOptionEntry instances which, when added to
a GOptionGroup, will set the values of the properties from the
parsed command line options.

Then, the above utility code is used to expose the properties from
WebKitSettings in a generic way, instead of having to manually add
a GOptionEntry and handling for each settings.

The following CLI options are replaced by the new functionality,
which covers the corresponding settings:

  --webgl          becomes --enable-webgl=1
  -D/--dev-tools   becomes --enable-developer-extras=1
  -v/--log-console becomes --enable-write-console-messages-to-stdout=1

Closes #27