CLI option for setting the policy for accepting cookies
This adds support for a new GOptionGroup which will contain all the
cookies-related command line flags, and adds a new --cookies=MODE
option which allows setting how cookies are accepted.

Some reorganization was needed to allow early creation of the
WebKitWebContext instance, which we need at the time when the option
group is added to the GApplication in order to pass the context's
WebKitCookieManager to the callback used to parse the command line:

- Move instantiation of the WebKitWebContext into the construction of
  the CogLauncher instance.
- Explicitly call g_set_prgname() instead of waiting for GApplication
  to do that, which is too late.
- Remove the cog_launcher_create_web_context(), which would have only
  the late registration of request handlers. The loop used for handler
  registration is moved into cog_launcher_startup().

Subsequent commmits will add additional entries to the GOptionGroup
used for cookie-related command line options.