CLI option for pre-setting cookies
This adds a --cookie-add command line option which can be used to
prefill the cookie jar with any number of cookies. This can be very
useful e.g. to provide authentication obtained after logging-in to
a Web service beforehand (probably by means other than a Web login
form), by passing the session cookie to Cog.