CLI option for configuring persistent cookie storage
Add a --cookie-jar=FORMAT[:PATH] which can be used to specify that
cookies which are valid for time longer than the session are to be
persisted in a certain FORMAT (either "plain" or "sqlite") at an
optional PATH. When not specified, the latter is derived from the
base data path of the WebKitWebsiteDataManager, resulting in the
cookie jar file being stored under the same directory as e.g. the
local storage and other related files.

Note that this neatly uses GType to introspect which enumeration
values are available to choose the FORMAT.