Reimplement CogLauncher in terms of CogShell
This removs from CogLauncher the functionality that was movd to
CogShell, and makes the former use CogShell. This is a straight
replacement: things work as they did before, without any forther
refinements. Down the like it will be needed to make make platform
plug-ins receive a CogShell instance, and adding plug-in management
to CogShell, for example.