core: Remove the CogShell::home-uri property
The default value for the property was NULL, which caused a critical
warning to be logged if it was not set to some value. At the same time,
setting a value would cause the URI to be loaded right away after the
Web view was created, which would add it to the back-forward navigation
list. Some applications may prefer to not have any entry pre-added to
the back-forward list, and anyway it's easy to just connect to the
CogShell::create-view signal (or chain up if subclassing) and do the
load there, so removing the property seems to be the best course of
action.