core: Make CogPlatform an interface, add CogPlatformView
This makes CogPlatform a GObject interface, which platform plug-ins
must implement. This makes platform implementations actual objects,
which can carry around the state they need instead of forcing the
implementations to use globals (which currently FDO does).

Creating a new view now returns an object which implements the new
CogPlatformView interface, instead of plainly returning an instance
of a WebKitWebViewBackend. In the same vein, this allows platform
implementations to keep the per-view state they may need in an object.
Platform plug-ins which do not need to keep per-view state can use
the new CogSimplePlatforView class.

The FDO platform plug-in is modified with the minimum set of changes
needed to implement the new interface, but it needs further work to
avoid globals and properly support multiple views.