Create Wayland event source after the WPE view backend
Certain Wayland events cause libwpe functions to be used, many of which
need a valid WPE view backend available. To avoid trying to use a NULL
view backend, this moves the creation of the GSource which integrates
Wayland event dispatch with the GLib main loop to be done always after
the first view backend is created.

This fixes a crash which was possible to trigger in the short window of
time between using cog_platform_setup() and cog_platform_get_view_backend()
e.g. by producing using input (e.g. pointer, key, or touch events) during
the time window.