core: When guess input URIs, directly use valid ones
Change the cog_uri_guess_from_user_input() function slighly, so if an
input URI is valid and the scheme refers to a network protocol, use the
passed string directly, decoded from the command line locale. This
avoids converting the SoupURI resulting from parsing to a string, which
is what caused passwords in URIs to be stripped out.

Fixes #75