CMake: Make custom commands create directories
Ninja will create itself the directories needed by the output files
for targets, but Make does not do that, nor does the CMake Makefiles
generator. Therefore, this adds a call to CMake's "make_directory"
tool to ensure that the output directory will always be available,
regardless of the build system being used.

Fixes #99