fdo: Use device scale API only when available in libwpe
Use wpe_view_backend_dispatch_set_device_scale_factor() only when
compiling against a version of libwpe equal or newer than 1.3.0,
which was the one introducing the new API.

The wl_data.current_output.scale member is set to "1" by default to
avoid needing to guard every single usage of it. The value is set
with a C99 designated initializer, which guarantees that the rest
of struct members are zeroed and allows removing a call to memset().