cog: Add --device-scale command line option
Add a "--device-scale" command line option which sets the initial
scaling factor for the output device. This is different from zooming
the Web content with "--scale", and is directly related to the pixel
density of the display, where a scaling factor of 1.0 corresponds to
the traditional the de-facto standard 96 DPI pixel density used by
most computer displays. If the DPI for a display is known, it can be
divided by 96 to find out a suitable device scaling factor for it.

Note that this sets only an initial value for the output device scale:
a platform plug-in can change this factor later on by itself. For example
the FDO platform plug-in will update it whenever the Wayland compositor
reports a change of output device.