pointer_on_axis: fix the axis-scrolling issues
The pointer_on_axis callback can receive a floating-point value that
represents the current vector distance of movement along that axis.
Current libwpe API cannot accomodate that value, so instead we should
only take the sign of that value and accordingly pass -1 or 1 as the
scrolling value for the axis event.