CI: Cache toolchain instead of always downloading it
This instructs Travis-CI to cache the contents of the ~/toolchain/
directory.  The "install" step is modified to use of "curl --time-cond"
passing ~/toolchain/.installed as the time reference, which makes curl
use the "If-Modified-Since" HTTP header to avoid re-downloading the
toolchain installer if the already installed toolchain (from the cache)
is as new as the installer file hosted by the server.