CI: Add WebKitGTK build to the matrix
This adds a new build to the CI configuration matrix which will use
Cog's WebKitGTK mode (instead of WPE) in release mode. A few shell
conditionals are used to avoid the toolchain download in this case,
in favor of a native build using the WebKitGTK distribution package.
The build matrix had to be expanded manually in order to allow
installation of different package sets using "addons.apt" depending
on the build job.

The version of CMake in Ubuntu Xenial does not support the -S/-B command
line flags, so instead manually create and use the _build/ subdirectory.