Turn device scaling factor into a CogShell property
This allows platform plug-ins reading the value of the property during
their initialization, and shifts to them the responsibility of applying
the device scaling factor in which way and moment is better for the
pltform plug-in.

The FDO platform plug-in does not need any changes, as it picks the
scaling factor from the Wayland compositor. The experimental DRM plug-in
is changed to take into account the value of the CogShell::device-scale
during view backend creation.