shell: Avoid G_ADD_PRIVATE warning
Building with GLib 2.62 produces a bunch of the following warnings:

  ../core/cog-shell.c: In function ‘cog_shell_startup_base’:
  ../core/cog-shell.c:114:13: warning: G_ADD_PRIVATE
    114 |     CogShellPrivate *priv = PRIV (shell);
        |                           ^~~~~~~~~~~~~~~

This uses cog_shell_get_instance_private(), which is defined
automatically by the G_DEFINE_TYPE_* macros and is the preferred
way of obtaining a pointer to the instance private area anyway.
GLib 2.4 or newer is needed, which was released in 2004, so this
is not expected to introduce any issue.

(cherry picked from commit 37e742b3253f032a9b2859179cd53710f1837022)