drm: cache imported gbm_bo objects and corresponding DRM framebuffers
For each exported wl_resource, we now cache the imported gbm_bo object and the
corresponding framebuffer object, avoiding constant cycling of these resources.

Upon first export, we spawn all the relevant resources. On further exports, we
iterate over the buffer list to find the cached object, and reuse such an object
if it exists.

Each cached buffer is associated to the wl_resource lifetime through the
destroy listener that's set up on the resource object. The cached buffers are
cleaned up during teardown.