cog-fdo-shell: Remove WPE-related globals and use callback data instead
The relationship between WPE and Wayland is n:1, hence having
globals for WPE-related concepts is problematic. Also, using
the list of exportables and browsing it every single time
when needing to decide on which exportable or backend to work on
is not efficient.

Use instead the data parameter in Wayland callback methods,
refactoring things wherever needed in order to be able to pass
the appropriate data: for Wayland callbacks we want in general
to pass the shell itself, so that later the callbacks can peak
and decide which view is the one we are currently on. For WPE
callbacks (that usually are connected when dealing directly
with a view's exportable or backend, for example) use a new
ViewBackendData structure, which we create for each new view
and stores the appropriate WPE data structures for each view's
backend.