pwl: Introduce pwl_window_{create,destroy} functions
Add the pwl_window_create() and pwl_window_destroy() functions.
This makes it possible to remove destroy_window() from pwl.h and
it's made static. Some of the base initialization code is moved
from create_window() to pwl_window_create(), as well.

This paves the way for moving code for window initialization and
disposal to the new functions.