pwl: Remove init_wayland() function
Drop the init_wayland() function and perform all the needed
initialization in pwl_display_connect(), where it belongs.
There is some churn moving code around to make sure a few
callback functions for global listeners are defined before
they are used.

The PwlError enumeration is added, and its corresponding
PWL_ERROR macro in order to report Wayland-related errors
to the caller of pwl_display_connect(). The checks for the
needed Wayland interfaces are reported using this instead
of dropping the ball on the floor due to a failed assertion.