pwl: Allow setting the application identifier
Add the pwl_window_set_application_id() functiom to configure the
application identifier reported to the compositor to identify top
level surfaces which belong to the same application.

Usually all the top level surfaces will all share the same identifier,
so this also adds a pwl_display_set_default_application_id() which can
be used to assign an identifier which will be set to all the PwlWindow
instances created afterwards.

For the moment pwl_window_set_application_id() only works if the
compositor exposes the XDG shell (xdg_wm_base) or wl_shell, and it
is a no-op otherwise.

Note that this does not allow configuring the application identifier
as a property of CogShell, that will be handled in a separate change.
For now COG_DEFAULT_APPID is used.