pwl: Always try to apply appid/title on window creation
Move the calls to set the window's application identifier and title
during window creation so they are always tried regardless of the shell
being used. This is needed to fix setting the values when wl_shell is in
use, because previously the calls were being issued only when
xdg_wm_base was being used.