fdo: bind to wp_presentation interface, issue feedback requests on surface updates
The `presentation-time` protocol is built in order to be able to bind to the
wp_presentation interface, if exposed by the compositor. Upon each surface
update the presentation feedback is requested. Once the feedback indicates
the update was either presented or discarded, the wp_presentation_feedback
object is also destroyed.

We don't actually use this interface for anything right now, but binding
to it enables seeing more detailed presentation information when running
with debugging options that enable protocol message dumps.