pwl: Handle pointer events at the window level
Move handling of pointer events from PwlDisplay to PwlWindow. In Wayland
events are always local to a surface i.e. (x, y) coordinates relate to a
certain surface (in our case: a window) and not to the display itself.

For this change the PwlPointer event is moved into PwlWindow, and the
.enter callback in the pointer listener is used to track the surface
(that is: window) to which subsequent events belong. The PwlWindow is
set as user data for their corresponding Wayland surfaces to easily
track the surface->window relationship.