Support 2D scrolling via axis events
Have the wl_pointer.axis() event gather 2D delta information. If possible,
the data is gathered and dispatch on the wl_pointer.frame() event,
otherwise immediately inside wl_pointer.axis().

When dispatching axis events, if the new 2D axis API is available and if the
underlying WebKit version is recent enough, the axis change is dispatched
through the wpe_input_axis_2d_event structure.

If that's not the case, the 2D movement is mimicked by dispatching for each
changed axis.