fdo: Support single-window mode
Add a single-window mode to CogFdoShell, which is a construction-time
property. This uses the same Wayland surface for all views, displaying
only the focused one. In this operation mode, the PwlWindow used is
owned by CogFdoShell, and each CogFdoView tracks the last wl_buffer
provided by the WPE backend in order to allow showing it immediately
when switching the focused view, instead of using a buffer listener.

There is some code for the window event handlers which is almost a
duplicate of the existing functions for the multi-window mode, and it
might be possible to factor it out later on. The code for requesting
frame callbacks and attaching a wl_buffer to windows has been already
cleaned up and is shared by both modes.

The cog_shell_new{,_from_module} function are updated to allow passing
property names as varargs on construction, which can be used to set the
CogFdoShell.single-window property (or other properties for other shell
implementations).