pwl: Add support for the ivi-application protocol
Support the ivi-application protocol, adding two new pwl_window_set_id()
and pwl_window_get_id() functions. If the ivi-application protocol is
being used, it will set the identifier for the surface associated to the
window, otherwise it is a no-op.

Based on a patch originally from Claudio Saavedra <csaavedra@igalia.com>