fdo: Handle view "in-window" and "visible" properties
Add handling of the CogView.in-window and .visible properties. In single
window mode they all go in sync because the focused view will always be
the only one displayed and in a window; in multiple window mode views
will be always in a window but not necessarily visible (this case is
currently unhandled due to missing support in pwl.[hc] to know when a
surface is on an output).