fdo: Set opaque region if background has no alpha
Always set an opaque region covering the whole web view when the web
view's background color does not have alpha, as WebKit renders frames
which are opaque anyway in that case. This makes the FDO shell set the
opaque region by default even for non-fullscreened surfaces, which
results in increased performance in many hardware configurations because
the compositor can skip a blending operation for the web view's window
surface.