fdo: Load and set a default pointer cursor
Load the default cursor theme and assign a surface with the left arrow
image as cursor for the seat's pointer. This is a basic implementation
which uses always the same cursor, even in HiDPI mode, so in that case
the cursor will be tiny.

Fixes #191

(cherry picked from commit 5fdf74513d7f0b6db4a4aacfc2a093e81de0831e)