pwl: Rework event source to avoid read lockups
Ensure that wl_display_prepare_read() is not used again after it has
been called once, by introducing a "reading" flag in the event source
which is set in the interval between preparing for reading and the
moment when it's either completed with wl_display_read_events() or
wl_display_cancel_read() -- at which point the flag is reset.

This is needed to avoid lockups inside the Wayland client library: as
per the documentation, calls to wl_display_prepare_read() have to be
balanced with actually completing the operation, either by reading the
events with wl_display_read_events() or canceling the read operation
with wl_display_cancel_read(). Not doing the sequence correctly can
lead to lockups because the Wayland client library will try to lock
the same resources more than once from the same thread, resulting in
a deadlock.

While at it, apply a number of cleanups: make naming coherent
(PwlSource, pwl_source_*), add factory function (pwl_source_new),
and replace setup_wayland_event_source() in favor of
pwl_display_attach_sources().