fdo: Support painting surface contents using OpenGL ES
Add a surface update method which uses Wayland-EGL + GLESv2 to scan out
frames provided by WPEBackend-fdo by painting a quad that covers the
whole viewport with the frame as its texture.

This can be used to workaround buggy drivers or drivers which have
partial Wayland support, where eglCreateWaylandBufferFromImageWL()
is unavailable or broken.