fdo: Support building against WPEBackend-fdo 1.2
Lower the requirement on WPEBackend-fdo so it is possible to build the
FDO shell module against the 1.2 series. This is achieved using version
check macros and falling back to the .export_egl_image callback instead
of using .export_fdo_egl_image if needed. To iron out differences, a few
helper functions are defined conditionally, avoiding the need for many
preprocessor conditionals all around the code.