shell: Do not add one ref too much on created views
Do not add an additional ref to the CogView instances when adding them
to the list of views managed by CogShell. This fixes disposal of views,
which was not happening on view removal because the additional ref was
preventing destruction.

In order to make it explicit that the shell_connect_view() internal
function takes ownership of the initial reference right after creation
of a CogView, pass it along with g_steal_pointer().

Additionally, as it is not expected that user code keeps extra refs to
CogView instances when it calls cog_shell_remove_view(), add a check
to validate that the CogView is indeed destroyed and emit a critical
warning if that is not the case.