fdo: Improve exported image logic and frame ACKs
Make CogFdoView instances own exported images and always keep the
current one around, to be released when the next one arrives. This
allows to simplify the wl_buffer listener to plainly destroy buffers.

Frames will be acknowledged always once they get to be displayed, which
means that in single window mode frame compositing will be throttled
inside WebKit and stopped until the view is in focus again. Note that
this is not a problem: things which do not depend on compositing
(basically everything except requestAnimationFrame) will keep working
from the point of view of web content.

Note that with this frame ACK policy, changes in web content that
require repainting or relayout are accumulated by WebKit, and once
the delayed frame is acknowledged they will be applied all at once and
if needed a new frame image exported: this can cause one-frame jolt if
the saved frame and the new one with accumulated changes are visually
different. A possible solution (not implemented by this patch) would be
to ACK the frame and start a short timer before using the old exported
image, and if a new exported image arrives while the timer is running
discard the old image and use the new one instead.

Lastly, cog_fdo_view_on_notify_focused() is fixed to correctly update
the surface with the last exported image when RENDER_MODE_GLES_V2_PAINT
is in use: previously re-focusing a view would cause wl_surface_attach()
to be called instead of cog_fdo_shell_gles_paint() on a surface which
had been configured with an EGL surface for painting.