drm: optionally limit max video mode width/height
Limit max video mode width/height so the selection of best video mode is limited by these boundaries. It's useful, for example, on Raspberry Pi 4 that supports 4K. If I connect it to 4K screen, it will by default automatically select 4K mode but in some cases that's not what I want. When there's a lot of elements on the screen, CSS animations, etc., performance seems to suffer in 4K so I'd prefer to just force it to be max Full HD.