Add initial X11-oriented platform module
The module is very basic in its nature, resampling the exported buffers
into a separate GL scene and using a dumb repaint loop to present that
scene in an X11 window. libxcb is used as the library through which
the X11 protocol is handled.

Very basic input (mouse, keyboard) is supported. The X11 window is
closable, but not (yet) through any keyboard shortcut.

This is intended as a minimal X11-functionality baseline, with
additional attention provided as required.