drm: introduce more debugging output in init_drm
During DRM initialization, debug output now lists as much
information as possible about possible devices, connectors and
modes that are being enumerated and chosen.

This would probably be helpful for anyone debugging the DRM
platform plugin functionality, and will be of further help
when additional configuration options are made available
for more fine-grained selection of devices, connectors, modes
and other attributes in the future.