drm: add support for atomic modesetting (#239)
Default to atomic modesetting, if the capability is detected as usable.
It can be force-disabled through the 'disable-atomic-modesetting=true'
configuration setting under the 'drm' group in the provided config
key-value file.

Additionally to connector and CRTC objects, we now also search through
the planes, picking the primary plane as desired for presentation.
Necessary property information for all these objects is also gathered.

When committing the framebuffer object, in case of using atomic
modesetting, an atomic request is built, enhanced with all the relevant
properties. Once committed, a page flip event is expected and handled
the same way as with the non-atomic modesetting.