Remove the support for building against WebKitGTK
Remove WebKitGTK-related code and references to the possibility of
building Cog using WebKitGTK instead of WPE WebKit.

The rationale for the removal is that the code path is not often used,
it increases maintenance effort slightly, the rework for multiple views
support (issue #77) will be more complicated if we need to maintain the
WebKitGTK support, and finally it is nowadays affordable to run and test
Cog with WPE directly on desktop than when the project was started (e.g.
using the X11 platform module, or in a nested Wayland compositor).