fdo: support wpe_fdo_shm_exported_buffer exports
When software-only GL rasterization is enforced (or when falling back
to it for whatever reason), the Wayland EGL platform implementation
can retreat to producing SHM-based buffers. When using the existing
wpe_view_backend_exportable_fdo_egl_client interface, these arrive as
wpe_fdo_shm_exported_buffer objects. This changeset introduces proper
handling for these types of exports.

This functionality is fairly new in WPEBackend-fdo, so the relevant
code is (where necessary) guarded with the HAVE_SHM_EXPORTED_BUFFER
macro that's enabled for the post-1.8.0 release series.

Once exported, we operate on the wl_resource and wl_shm_buffer objects
that relevantly describe this buffer. For any such export, we provide
a separate shared-memory region that is basis for its own wl_shm_pool
and the wl_buffer that's retrieved from that pool.

We rely on the underlying graphics driver to efficiently handle these
buffers so that our own mappings can be reused as long as the exported
buffers don't change too frequently. This allows us to cache our own
buffer objects based on the exports we receive.

Once we have our own memory regions set up or retrieved from the
cache, we copy over the contents from the exported buffer to the
buffer region we maintain, and the wl_buffer object based on that
region is pushed to the compositor. Right now we effectively mirror
the exported SHM buffer, replicating the size, stride and format of
that buffer.

Buffer objects are cached in a global wl_list. For each export, that
list is traversed to find anything reusable. Each such new cache entry
is also removed when the underlying wl_resource of the export is
destroyed or upon termination.