drm: support wpe_fdo_shm_exported_buffer exports
Handle the wpe_fdo_shm_exported_buffer exports that might arrive on
the wpe_view_backend_exportable_fdo_client interface.

When that occurs, we craft a gbm_bo with the necessary flags that
would allow us to map the bo's underlying memory and write into it.
We still create a framebuffer object off of it, pretty much like we
do it for other types of bos.

The new buffer objects use the existing caching facilities. Each
buffer object, cached or newly created, is updated by mapping the bo
memory into the address space and copying over the data contained
in the shm area.

For now, only ARGB8888 and XRGB8888 SHM formats are handled. In both
cases the corresponding gbm bo will use the XRGB8888 format.