core: Add cog-host-routes-handler
Handler for custom URI scheme requests that can route requests to
different handlers depending on the *host* component of the requested
URI.

Optionally, if a “fallback” handler has been specified, it will be
used to serve requests which do not match any of the routed hosts.

The set of available host map entries can be configured using
[method@Cog.HostMapHandler.add] and
[method@Cog.HostMapHandler.delete]. For each request, map entries
are checked and the one that matches the URI *host* component
will handle the request.

This handler is typically used in tandem with
[class@Cog.DirectoryFilesHandler], the latter being typically a
fallback, or as the handler for a routed host.

Signed-off-by: Pablo Saavedra <psaavedra@igalia.com>