fdo: Reintroduce checks on Wayland protocols
Add code to do a cursory check on the Wayland protocols handled by the
Wayland compositor in the check_supported() function. This makes sure
that the FDO platform will be chosen if at least one of the shell
protocols supported by the code are handled by the compositor.

Note that it is not needed to check for protocols from the base Wayland
specification which every compositor is expected to implement
(wl_compositor, wl_shm, wl_seat, etc.) because these will always be
available.

As a related note, it would be neat to check that the EGL will be usable
with the Wayland compositor, but there is no 100% reliable way without
going through actual EGL initialization, so no check is added for this
at the moment.